Break up abstract and description
The abstract is a one-sentence summary of the description. It reads a
bit strange to see both of them combined together. This adds a line
between the two to emphasize the distinction.

Not sure how we plan on eventually displaying these visually on the
website. Will the abstract/description be separate? Are we trying to
combine the two into a single description?